RES-3466: relocate const_eval_ext regression tests
