+++ resilient/src/const_eval_ext.rs
@@ -143,19 +143,6 @@ mod tests {
         Node::Program(vec![stmt])
     }
 
-    fn program_with(stmts: Vec<Spanned<Node>>) -> Node {
-        Node::Program(stmts)
-    }
-
-    fn assert_check_ok(program: Node) {
-        assert!(check(&program, "test.rz").is_ok());
-    }
-
-    fn assert_check_err(program: Node, expected: &str, label: &str) {
-        let err = check(&program, "test.rz").unwrap_err();
-        assert_eq!(err, expected, "case {label}");
-    }
-
     #[test]
     fn const_string_concat() {
         let out = run(r#"
@@ -320,145 +307,5 @@ println(to_string(a + b));
             err,
             "test.rz:7:2: error: invalid const declaration: type annotations require an initializer"
         );
-    }
-
-    #[test]
-    fn const_decl_check_accepts_valid_baselines() {
-        assert_check_ok(program(const_stmt(
-            "ANSWER",
-            Node::IntegerLiteral {
-                value: 42,
-                span: span(1, 14),
-            },
-            None,
-            1,
-            1,
-        )));
-
-        assert_check_ok(program(const_stmt(
-            "LIMIT",
-            Node::IntegerLiteral {
-                value: 7,
-                span: span(4, 18),
-            },
-            Some("int"),
-            4,
-            3,
-        )));
-
-        assert_check_ok(program_with(vec![
-            const_stmt(
-                "ZERO",
-                Node::IntegerLiteral {
-                    value: 0,
-                    span: span(8, 16),
-                },
-                None,
-                8,
-                1,
-            ),
-            const_stmt(
-                "FOUR",
-                Node::IntegerLiteral {
-                    value: 4,
-                    span: span(9, 15),
-                },
-                None,
-                9,
-                1,
-            ),
-        ]));
-    }
-
-    #[test]
-    fn const_decl_check_rejects_malformed_regressions() {
-        let cases = [
-            (
-                "missing name",
-                program(const_stmt(
-                    "",
-                    Node::IntegerLiteral {
-                        value: 1,
-                        span: span(1, 18),
-                    },
-                    None,
-                    1,
-                    1,
-                )),
-                "test.rz:1:1: error: invalid const declaration: missing name",
-            ),
-            (
-                "whitespace-only name",
-                program(const_stmt(
-                    " \t",
-                    Node::IntegerLiteral {
-                        value: 2,
-                        span: span(2, 17),
-                    },
-                    None,
-                    2,
-                    4,
-                )),
-                "test.rz:2:4: error: invalid const declaration: missing name",
-            ),
-            (
-                "missing initializer",
-                program(const_stmt(
-                    "COUNT",
-                    Node::IntegerLiteral {
-                        value: 0,
-                        span: Span::default(),
-                    },
-                    None,
-                    3,
-                    5,
-                )),
-                "test.rz:3:5: error: invalid const declaration: missing initializer",
-            ),
-            (
-                "typed missing initializer",
-                program(const_stmt(
-                    "VALUE",
-                    Node::IntegerLiteral {
-                        value: 0,
-                        span: Span::default(),
-                    },
-                    Some("int"),
-                    7,
-                    2,
-                )),
-                "test.rz:7:2: error: invalid const declaration: type annotations require an initializer",
-            ),
-            (
-                "duplicate malformed forms",
-                program_with(vec![
-                    const_stmt(
-                        " ",
-                        Node::IntegerLiteral {
-                            value: 0,
-                            span: Span::default(),
-                        },
-                        None,
-                        11,
-                        3,
-                    ),
-                    const_stmt(
-                        " ",
-                        Node::IntegerLiteral {
-                            value: 0,
-                            span: Span::default(),
-                        },
-                        None,
-                        14,
-                        1,
-                    ),
-                ]),
-                "test.rz:11:3: error: invalid const declaration: missing name",
-            ),
-        ];
-
-        for (label, program, expected) in cases {
-            assert_check_err(program, expected, label);
-        }
     }
 }
