+++ resilient/src/lib.rs
@@ -683,8 +683,6 @@ mod unused_imports;
 mod tuple_struct;
 // RES-2580: extended const eval — string concat, bitwise ops, conditionals.
 mod const_eval_ext;
-#[cfg(test)]
-mod const_eval_ext_regression;
 // RES-2601: exhaustive struct field checking in match patterns.
 mod struct_field_check;
 // RES-2586: deque (double-ended queue) collection builtins.
